JAMES-2214 Rename ambiguous and misleading MailboxNotImplementedException
  MailboxNotImplementedException and the "Not yet implemented" message is misleading.

  JMAP spec does not support, whatever version you take, saving messages in other mailboxes than Outbox or Draft.

  The error message name should explain that the given mailboxes were invalid. Moreover, the error message should explain which state was not met by the provided mailboxes.
